Switch left ctrl and caps lock to match with WM
I switched them in hyprland to have better access to ctrl when typing
directly on the keyboard of the laptop, but then I need to switch them
in here as well so that my ctrl keys remain in the same place.

Also, it's not very practical for certain things to not have ctrl in
direct access, like to go back and forth in the jump list in helix
(ctrl+o/i).